pyproject.toml
[tool.poetry]
name = "kalalogs"
version = "0.1.0"
description = "Simple colored logging in Python"
authors = ["Junyi <47441677+immanuelhume@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
